requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.5.0
idna==2.10
iniconfig==1.1.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.1.2
pytest-django==4.1.0
pytz==2020.4
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.11
gunicorn==20.0.4
psycopg2-binary==2.8.5